Extract Contracts component from ContractsApp
